// src/app/model/xivapi-item.model.ts
export interface XivapiItemModel {
  Adjective: number;
  AetherialReduce: number;
  AlwaysCollectable: number;
  Article: number;
  BaseParam0: BaseParam0;
  BaseParam0Target: string;
  BaseParam0TargetID: number;
  BaseParam1: BaseParam1;
  BaseParam1Target: string;
  BaseParam1TargetID: number;
  BaseParam2: BaseParam2;
  BaseParam2Target: string;
  BaseParam2TargetID: number;
  BaseParam3: BaseParam3;
  BaseParam3Target: string;
  BaseParam3TargetID: number;
  BaseParam4: any;
  BaseParam4Target: string;
  BaseParam4TargetID: number;
  BaseParam5: any;
  BaseParam5Target: string;
  BaseParam5TargetID: number;
  BaseParamModifier: number;
  BaseParamSpecial0: any;
  BaseParamSpecial0Target: string;
  BaseParamSpecial0TargetID: number;
  BaseParamSpecial1: any;
  BaseParamSpecial1Target: string;
  BaseParamSpecial1TargetID: number;
  BaseParamSpecial2: any;
  BaseParamSpecial2Target: string;
  BaseParamSpecial2TargetID: number;
  BaseParamSpecial3: any;
  BaseParamSpecial3Target: string;
  BaseParamSpecial3TargetID: number;
  BaseParamSpecial4: any;
  BaseParamSpecial4Target: string;
  BaseParamSpecial4TargetID: number;
  BaseParamSpecial5: any;
  BaseParamSpecial5Target: string;
  BaseParamSpecial5TargetID: number;
  BaseParamValue0: number;
  BaseParamValue1: number;
  BaseParamValue2: number;
  BaseParamValue3: number;
  BaseParamValue4: number;
  BaseParamValue5: number;
  BaseParamValueSpecial0: number;
  BaseParamValueSpecial1: number;
  BaseParamValueSpecial2: number;
  BaseParamValueSpecial3: number;
  BaseParamValueSpecial4: number;
  BaseParamValueSpecial5: number;
  Block: number;
  BlockRate: number;
  CanBeHq: number;
  CastTimeS: number;
  ClassJobCategory: ClassJobCategory;
  ClassJobCategoryTarget: string;
  ClassJobCategoryTargetID: number;
  ClassJobRepair: ClassJobRepair;
  ClassJobRepairTarget: string;
  ClassJobRepairTargetID: number;
  ClassJobUseTarget: string;
  ClassJobUseTargetID: number;
  CooldownS: number;
  DamageMag: number;
  DamagePhys: number;
  DefenseMag: number;
  DefensePhys: number;
  DelayMs: number;
  Description: string;
  DescriptionJSON_cn: any;
  DescriptionJSON_kr: any;
  Description_cn: any;
  Description_de: string;
  Description_en: string;
  Description_fr: string;
  Description_ja: string;
  Description_kr: any;
  Desynth: number;
  EquipRestriction: number;
  EquipSlotCategory: EquipSlotCategory2;
  EquipSlotCategoryTarget: string;
  EquipSlotCategoryTargetID: number;
  FilterGroup: number;
  GameContentLinks: any;
  GamePatch: GamePatch;
  GrandCompany: any;
  GrandCompanyTarget: string;
  GrandCompanyTargetID: number;
  ID: number;
  Icon: string;
  IconHD: string;
  IconID: number;
  IsAdvancedMeldingPermitted: number;
  IsCollectable: number;
  IsCrestWorthy: number;
  IsDyeable: number;
  IsGlamourous: number;
  IsIndisposable: number;
  IsPvP: number;
  IsUnique: number;
  IsUntradable: number;
  ItemAction: any;
  ItemActionTarget: string;
  ItemActionTargetID: number;
  ItemGlamour: ItemGlamour;
  ItemGlamourTarget: string;
  ItemGlamourTargetID: number;
  ItemKind: ItemKind;
  ItemRepair: ItemRepair;
  ItemRepairTarget: string;
  ItemRepairTargetID: number;
  ItemSearchCategory: any;
  ItemSearchCategoryTarget: string;
  ItemSearchCategoryTargetID: number;
  ItemSeries: any;
  ItemSeriesTarget: string;
  ItemSeriesTargetID: number;
  ItemSortCategory: ItemSortCategory3;
  ItemSortCategoryTarget: string;
  ItemSortCategoryTargetID: number;
  ItemSpecialBonus: any;
  ItemSpecialBonusParam: number;
  ItemSpecialBonusTarget: string;
  ItemSpecialBonusTargetID: number;
  ItemUICategory: ItemUicategory3;
  ItemUICategoryTarget: string;
  ItemUICategoryTargetID: number;
  LevelEquip: number;
  LevelItem: number;
  Lot: number;
  Materia: any;
  MateriaSlotCount: number;
  MaterializeType: number;
  ModelMain: string;
  ModelSub: string;
  Name: string;
  NameCombined: string;
  NameCombined_cn: string;
  NameCombined_de: string;
  NameCombined_en: string;
  NameCombined_fr: string;
  NameCombined_ja: string;
  NameCombined_kr: string;
  NameLocale: string;
  Name_de: string;
  Name_en: string;
  Name_fr: string;
  Name_ja: string;
  Patch: number;
  Plural: string;
  Plural_de: string;
  Plural_en: string;
  Plural_fr: string;
  Plural_ja: string;
  PossessivePronoun: number;
  PriceLow: number;
  PriceMid: number;
  Pronoun: number;
  Rarity: number;
  Singular: string;
  Singular_de: string;
  Singular_en: string;
  Singular_fr: string;
  Singular_ja: string;
  StackSize: number;
  StartsWithVowel: number;
  Stats: Stats;
  SubStatCategory: number;
  Url: string;
  UrlType: string;
  _: string;
  _Score: number;
}

export interface BaseParam0 {
  '1HWpn%': number;
  '2HWpn%': number;
  'Bracelet%': number;
  'Chest%': number;
  'ChestHead%': number;
  'ChestHeadLegsFeet%': number;
  'ChestLegsFeet%': number;
  'ChestLegsGloves%': number;
  Description: string;
  Description_de: string;
  Description_en: string;
  Description_fr: string;
  Description_ja: string;
  'Earring%': number;
  'Feet%': number;
  'Hands%': number;
  'Head%': number;
  'HeadChestHandsLegsFeet%': number;
  ID: number;
  'Legs%': number;
  'LegsFeet%': number;
  MeldParam0: number;
  MeldParam1: number;
  MeldParam10: number;
  MeldParam11: number;
  MeldParam12: number;
  MeldParam2: number;
  MeldParam3: number;
  MeldParam4: number;
  MeldParam5: number;
  MeldParam6: number;
  MeldParam7: number;
  MeldParam8: number;
  MeldParam9: number;
  Name: string;
  Name_de: string;
  Name_en: string;
  Name_fr: string;
  Name_ja: string;
  'Necklace%': number;
  'OH%': number;
  OrderPriority: number;
  PacketIndex: number;
  'Ring%': number;
  'UnderArmor%': number;
  'Waist%': number;
}

export interface BaseParam1 {
  '1HWpn%': number;
  '2HWpn%': number;
  'Bracelet%': number;
  'Chest%': number;
  'ChestHead%': number;
  'ChestHeadLegsFeet%': number;
  'ChestLegsFeet%': number;
  'ChestLegsGloves%': number;
  Description: string;
  Description_de: string;
  Description_en: string;
  Description_fr: string;
  Description_ja: string;
  'Earring%': number;
  'Feet%': number;
  'Hands%': number;
  'Head%': number;
  'HeadChestHandsLegsFeet%': number;
  ID: number;
  'Legs%': number;
  'LegsFeet%': number;
  MeldParam0: number;
  MeldParam1: number;
  MeldParam10: number;
  MeldParam11: number;
  MeldParam12: number;
  MeldParam2: number;
  MeldParam3: number;
  MeldParam4: number;
  MeldParam5: number;
  MeldParam6: number;
  MeldParam7: number;
  MeldParam8: number;
  MeldParam9: number;
  Name: string;
  Name_de: string;
  Name_en: string;
  Name_fr: string;
  Name_ja: string;
  'Necklace%': number;
  'OH%': number;
  OrderPriority: number;
  PacketIndex: number;
  'Ring%': number;
  'UnderArmor%': number;
  'Waist%': number;
}

export interface BaseParam2 {
  '1HWpn%': number;
  '2HWpn%': number;
  'Bracelet%': number;
  'Chest%': number;
  'ChestHead%': number;
  'ChestHeadLegsFeet%': number;
  'ChestLegsFeet%': number;
  'ChestLegsGloves%': number;
  Description: string;
  Description_de: string;
  Description_en: string;
  Description_fr: string;
  Description_ja: string;
  'Earring%': number;
  'Feet%': number;
  'Hands%': number;
  'Head%': number;
  'HeadChestHandsLegsFeet%': number;
  ID: number;
  'Legs%': number;
  'LegsFeet%': number;
  MeldParam0: number;
  MeldParam1: number;
  MeldParam10: number;
  MeldParam11: number;
  MeldParam12: number;
  MeldParam2: number;
  MeldParam3: number;
  MeldParam4: number;
  MeldParam5: number;
  MeldParam6: number;
  MeldParam7: number;
  MeldParam8: number;
  MeldParam9: number;
  Name: string;
  Name_de: string;
  Name_en: string;
  Name_fr: string;
  Name_ja: string;
  'Necklace%': number;
  'OH%': number;
  OrderPriority: number;
  PacketIndex: number;
  'Ring%': number;
  'UnderArmor%': number;
  'Waist%': number;
}

export interface BaseParam3 {
  '1HWpn%': number;
  '2HWpn%': number;
  'Bracelet%': number;
  'Chest%': number;
  'ChestHead%': number;
  'ChestHeadLegsFeet%': number;
  'ChestLegsFeet%': number;
  'ChestLegsGloves%': number;
  Description: string;
  Description_de: string;
  Description_en: string;
  Description_fr: string;
  Description_ja: string;
  'Earring%': number;
  'Feet%': number;
  'Hands%': number;
  'Head%': number;
  'HeadChestHandsLegsFeet%': number;
  ID: number;
  'Legs%': number;
  'LegsFeet%': number;
  MeldParam0: number;
  MeldParam1: number;
  MeldParam10: number;
  MeldParam11: number;
  MeldParam12: number;
  MeldParam2: number;
  MeldParam3: number;
  MeldParam4: number;
  MeldParam5: number;
  MeldParam6: number;
  MeldParam7: number;
  MeldParam8: number;
  MeldParam9: number;
  Name: string;
  Name_de: string;
  Name_en: string;
  Name_fr: string;
  Name_ja: string;
  'Necklace%': number;
  'OH%': number;
  OrderPriority: number;
  PacketIndex: number;
  'Ring%': number;
  'UnderArmor%': number;
  'Waist%': number;
}

export interface ClassJobCategory {
  ACN: number;
  ADV: number;
  ALC: number;
  ARC: number;
  ARM: number;
  AST: number;
  BLM: number;
  BLU: number;
  BRD: number;
  BSM: number;
  BTN: number;
  CNJ: number;
  CRP: number;
  CUL: number;
  DNC: number;
  DRG: number;
  DRK: number;
  FSH: number;
  GLA: number;
  GNB: number;
  GSM: number;
  ID: number;
  LNC: number;
  LTW: number;
  MCH: number;
  MIN: number;
  MNK: number;
  MRD: number;
  NIN: number;
  Name: string;
  Name_de: string;
  Name_en: string;
  Name_fr: string;
  Name_ja: string;
  PGL: number;
  PLD: number;
  RDM: number;
  ROG: number;
  RPR: number;
  SAM: number;
  SCH: number;
  SGE: number;
  SMN: number;
  THM: number;
  WAR: number;
  WHM: number;
  WVR: number;
}

export interface ClassJobRepair {
  Abbreviation: string;
  Abbreviation_de: string;
  Abbreviation_en: string;
  Abbreviation_fr: string;
  Abbreviation_ja: string;
  BattleClassIndex: string;
  CanQueueForDuty: number;
  ClassJobCategory: ClassJobCategory2;
  ClassJobCategoryTarget: string;
  ClassJobCategoryTargetID: number;
  ClassJobParent: any;
  ClassJobParentTarget: string;
  ClassJobParentTargetID: number;
  DohDolJobIndex: number;
  ExpArrayIndex: number;
  ID: number;
  Icon: string;
  IsLimitedJob: number;
  ItemSoulCrystal: ItemSoulCrystal;
  ItemSoulCrystalTarget: string;
  ItemSoulCrystalTargetID: number;
  ItemStartingWeapon: any;
  ItemStartingWeaponTarget: string;
  ItemStartingWeaponTargetID: number;
  JobIndex: number;
  LimitBreak1: any;
  LimitBreak1Target: string;
  LimitBreak1TargetID: number;
  LimitBreak2: any;
  LimitBreak2Target: string;
  LimitBreak2TargetID: number;
  LimitBreak3: any;
  LimitBreak3Target: string;
  LimitBreak3TargetID: number;
  ModifierDexterity: number;
  ModifierHitPoints: number;
  ModifierIntelligence: number;
  ModifierManaPoints: number;
  ModifierMind: number;
  ModifierPiety: number;
  ModifierStrength: number;
  ModifierVitality: number;
  MonsterNote: any;
  MonsterNoteTarget: string;
  MonsterNoteTargetID: number;
  Name: string;
  NameEnglish: string;
  NameEnglish_de: string;
  NameEnglish_en: string;
  NameEnglish_fr: string;
  NameEnglish_ja: string;
  Name_de: string;
  Name_en: string;
  Name_fr: string;
  Name_ja: string;
  PartyBonus: number;
  Prerequisite: any;
  PrerequisiteTarget: string;
  PrerequisiteTargetID: number;
  PrimaryStat: number;
  PvPActionSortRow: number;
  RelicQuest: any;
  RelicQuestTarget: string;
  RelicQuestTargetID: number;
  Role: number;
  StartingLevel: number;
  StartingTown: any;
  StartingTownTarget: string;
  StartingTownTargetID: number;
  UIPriority: number;
  UnlockQuest: any;
  UnlockQuestTarget: string;
  UnlockQuestTargetID: number;
}
// ...
export interface EquipSlotCategory {
  Body: number;
  Ears: number;
  Feet: number;
  FingerL: number;
  FingerR: number;
  Gloves: number;
  Head: number;
  ID: number;
  Legs: number;
  MainHand: number;
  Neck: number;
  OffHand: number;
  SoulCrystal: number;
  Waist: number;
  Wrists: number;
}