// src/app/model/xivapi-achievement.model.ts
export interface XviapiAchievementModel {
  AchievementCategory: AchievementCategory;
  AchievementCategoryTarget: string;
  AchievementCategoryTargetID: number;
  AchievementHideCondition?: AchievementHideCondition;
  AchievementHideConditionTarget: string;
  AchievementHideConditionTargetID: number;
  AchievementTarget?: AchievementTarget;
  AchievementTargetTarget: string;
  AchievementTargetTargetID: number;
  Data0: number;
  Data1: number;
  Data2: number;
  Data3: number;
  Data4: number;
  Data5: number;
  Data6: number;
  Data7: number;
  Description: string;
  Description_de: string;
  Description_en: string;
  Description_fr: string;
  Description_ja: string;
  GameContentLinks?: GameContentLinks;
  GamePatch: GamePatch;
  ID: number;
  Icon: string;
  IconHD: string;
  IconID: number;
  Item?: Item;
  ItemTarget: string;
  ItemTargetID: number;
  Key: number;
  Name: string;
  NameCombined: string;
  NameCombined_cn: string;
  NameCombined_de: string;
  NameCombined_en: string;
  NameCombined_fr: string;
  NameCombined_ja: string;
  NameCombined_kr: string;
  NameLocale: string;
  Name_de: string;
  Name_en: string;
  Name_fr: string;
  Name_ja: string;
  Order: number;
  Patch: number;
  Points: number;
  QuestRequirementsAll: boolean;
  Title?: Title;
  TitleTarget: string;
  TitleTargetID: number;
  Type: number;
  Url: string;
  UrlType: string;
  _: string;
  _Score: number;
}

export interface AchievementCategory {
  AchievementKind: AchievementKind;
  AchievementKindTarget: string;
  AchievementKindTargetID: number;
  HideCategory: number;
  ID: number;
  Name: string;
  Name_de: string;
  Name_en: string;
  Name_fr: string;
  Name_ja: string;
  Order: number;
  ShowComplete: number;
}

export interface AchievementKind {
  ID: number;
  Name: string;
  Name_de: string;
  Name_en: string;
  Name_fr: string;
  Name_ja: string;
  Order: number;
}

export interface AchievementHideCondition {
  HideAchievement: number;
  HideConditions: number;
  HideName: number;
  ID: number;
}

export interface AchievementTarget {
  ID: number;
  Type: number;
  Value: number;
}

export interface GameContentLinks {
  GilShopItem?: GilShopItem;
  SpecialShop?: SpecialShop;
  Achievement?: Achievement;
}

export interface GilShopItem {
  AchievementRequired: string[];
  RowRequired2?: string[];
}

export interface SpecialShop {
  AchievementUnlock4?: number[];
  AchievementUnlock3?: number[];
  AchievementUnlock2?: number[];
  AchievementUnlock1?: number[];
  AchievementUnlock0?: number[];
}

export interface Achievement {
  Data7?: number[];
  Data6?: number[];
  Key?: number[];
  Data0?: number[];
  Data1?: number[];
  Data2?: number[];
  Data3?: number[];
  Data4?: number[];
  Data5?: number[];
}

export interface GamePatch {
  Banner?: string;
  ExName: string;
  ExVersion: number;
  ID: number;
  Name: string;
  Name_cn: string;
  Name_de: string;
  Name_en: string;
  Name_fr: string;
  Name_ja: string;
  Name_kr: string;
  ReleaseDate: number;
  Version: string;
  PatchNotes?: string;
  PatchNotes_de?: string;
  PatchNotes_en?: string;
  PatchNotes_fr?: string;
  PatchNotes_ja?: string;
  IsExpansion?: boolean;
  Url?: string;
}

export interface Item {
  AdditionalData: number;
  Adjective: number;
  AetherialReduce: number;
  AlwaysCollectable: number;
  Article: number;
  BaseParam0?: BaseParam0;
  BaseParam0Target: string;
  BaseParam0TargetID: number;
  BaseParam1: any;
  BaseParam1Target: string;
  BaseParam1TargetID: number;
  BaseParam2: any;
  BaseParam2Target: string;
  BaseParam2TargetID: number;
  BaseParam3: any;
  BaseParam3Target: string;
  BaseParam3TargetID: number;
  BaseParam4: any;
  BaseParam4Target: string;
  BaseParam4TargetID: number;
  BaseParam5: any;
  BaseParam5Target: string;
  BaseParam5TargetID: number;
  BaseParamModifier: number;
  BaseParamSpecial0: any;
  BaseParamSpecial0Target: string;
  BaseParamSpecial0TargetID: number;
  BaseParamSpecial1: any;
  BaseParamSpecial1Target: string;
  BaseParamSpecial1TargetID: number;
  BaseParamSpecial2: any;
  BaseParamSpecial2Target: string;
  BaseParamSpecial2TargetID: number;
  BaseParamSpecial3: any;
  BaseParamSpecial3Target: string;
  BaseParamSpecial3TargetID: number;
  BaseParamSpecial4: any;
  BaseParamSpecial4Target: string;
  BaseParamSpecial4TargetID: number;
  BaseParamSpecial5: any;
  BaseParamSpecial5Target: string;
  BaseParamSpecial5TargetID: number;
  BaseParamValue0: number;
  BaseParamValue1: number;
  BaseParamValue2: number;
  BaseParamValue3: number;
  BaseParamValue4: number;
  BaseParamValue5: number;
  BaseParamValueSpecial0: number;
  BaseParamValueSpecial1: number;
  BaseParamValueSpecial2: number;
  BaseParamValueSpecial3: number;
  BaseParamValueSpecial4: number;
  BaseParamValueSpecial5: number;
  Block: number;
  BlockRate: number;
  CanBeHq: number;
  CastTimeS: number;
  ClassJobCategory?: ClassJobCategory;
  ClassJobCategoryTarget: string;
  ClassJobCategoryTargetID: number;
  ClassJobRepair?: ClassJobRepair;
  ClassJobRepairTarget: string;
  ClassJobRepairTargetID: number;
  ClassJobUse: any;
  ClassJobUseTarget: string;
  ClassJobUseTargetID: number;
  CooldownS: number;
  DamageMag: number;
  DamagePhys: number;
  DefenseMag: number;
  DefensePhys: number;
  DelayMs: number;
  Description: string;
  Description_de: string;
  Description_en: string;
  Description_fr: string;
  Description_ja: string;
  Desynth: number;
  EquipRestriction: number;
  EquipSlotCategory?: EquipSlotCategory;
  EquipSlotCategoryTarget: string;
  EquipSlotCategoryTargetID: number;
  FilterGroup: number;
  GrandCompany: any;
  GrandCompanyTarget: string;
  GrandCompanyTargetID: number;
  ID: number;
  Icon: string;
  IconHD: string;
  IconID: number;
  IsAdvancedMeldingPermitted: number;
  IsCollectable: number;
  IsCrestWorthy: number;
  IsDyeable: number;
  IsGlamourous: number;
  IsIndisposable: number;
  IsPvP: number;
  IsUnique: number;
  IsUntradable: number;
  ItemAction?: ItemAction;
  ItemActionTarget: string;
  ItemActionTargetID: number;
  ItemGlamour?: ItemGlamour;
  ItemGlamourTarget: string;
  ItemGlamourTargetID: number;
  ItemRepair?: ItemRepair;
  ItemRepairTarget: string;
  ItemRepairTargetID: number;
  ItemSearchCategory: any;
  ItemSearchCategoryTarget: string;
  ItemSearchCategoryTargetID: number;
  ItemSeries: any;
  ItemSeriesTarget: string;
  ItemSeriesTargetID: number;
  ItemSortCategory: ItemSortCategory;
  ItemSortCategoryTarget: string;
  ItemSortCategoryTargetID: number;
  ItemSpecialBonus: any;
  ItemSpecialBonusParam: number;
  ItemSpecialBonusTarget: string;
  ItemSpecialBonusTargetID: number;
  ItemUICategory: ItemUicategory;
  ItemUICategoryTarget: string;
  ItemUICategoryTargetID: number;
  LevelEquip: number;
  LevelItem: number;
  Lot: number;
  MateriaSlotCount: number;
  MaterializeType: number;
  ModelMain: string;
  ModelSub: string;
  Name: string;
  Name_de: string;
  Name_en: string;
  Name_fr: string;
  Name_ja: string;
  Plural: string;
  Plural_de: string;
  Plural_en: string;
  Plural_fr: string;
  Plural_ja: string;
  PossessivePronoun: number;
  PriceLow: number;
  PriceMid: number;
  Pronoun: number;
  Rarity: number;
  Singular: string;
  Singular_de: string;
  Singular_en: string;
  Singular_fr: string;
  Singular_ja: string;
  StackSize: number;
  StartsWithVowel: number;
  SubStatCategory: number;
}

export interface BaseParam0 {
  '1HWpn%': number;
  '2HWpn%': number;
  'Bracelet%': number;
  'Chest%': number;
  'ChestHead%': number;
  'ChestHeadLegsFeet%': number;
  'ChestLegsFeet%': number;
  'ChestLegsGloves%': number;
  Description: string;
  Description_de: string;
  Description_en: string;
  Description_fr: string;
  Description_ja: string;
  'Earring%': number;
  'Feet%': number;
  'Hands%': number;
  'Head%': number;
  'HeadChestHandsLegsFeet%': number;
  ID: number;
  'Legs%': number;
  'LegsFeet%': number;
  MeldParam0: number;
  MeldParam1: number;
  MeldParam10: number;
  MeldParam11: number;
  MeldParam12: number;
  MeldParam2: number;
  MeldParam3: number;
  MeldParam4: number;
  MeldParam5: number;
  MeldParam6: number;
  MeldParam7: number;
  MeldParam8: number;
  MeldParam9: number;
  Name: string;
  Name_de: string;
  Name_en: string;
  Name_fr: string;
  Name_ja: string;
  'Necklace%': number;
  'OH%': number;
  OrderPriority: number;
  PacketIndex: string;
  'Ring%': number;
  'UnderArmor%': number;
  'Waist%': number;
}

export interface ClassJobCategory {
  ACN: number;
  ADV: number;
  ALC: number;
  ARC: number;
  ARM: number;
  AST: number;
  BLM: number;
  BLU: number;
  BRD: number;
  BSM: number;
  BTN: number;
  CNJ: number;
  CRP: number;
  CUL: number;
  DNC: number;
  DRG: number;
  DRK: number;
  FSH: number;
  GLA: number;
  GNB: number;
  GSM: number;
  ID: number;
  LNC: number;
  LTW: number;
  MCH: number;
  MIN: number;
  MNK: number;
  MRD: number;
  NIN: number;
  Name: string;
  Name_de: string;
  Name_en: string;
  Name_fr: string;
  Name_ja: string;
  PGL: number;
  PLD: number;
  RDM: number;
  ROG: number;
  RPR: number;
  SAM: number;
  SCH: number;
  SGE: number;
  SMN: number;
  THM: number;
  WAR: number;
  WHM: number;
  WVR: number;
}

export interface ClassJobRepair {
  Abbreviation: string;
  Abbreviation_de: string;
  Abbreviation_en: string;
  Abbreviation_fr: string;
  Abbreviation_ja: string;
  BattleClassIndex: string;
  CanQueueForDuty: number;
  ClassJobCategory: ClassJobCategory2;
  ClassJobCategoryTarget: string;
  ClassJobCategoryTargetID: number;
  ClassJobParent: any;
  ClassJobParentTarget: string;
  ClassJobParentTargetID: number;
  DohDolJobIndex: number;
  ExpArrayIndex: number;
  ID: number;
  Icon: string;
  IsLimitedJob: number;
  ItemSoulCrystal: ItemSoulCrystal;
  ItemSoulCrystalTarget: string;
  ItemSoulCrystalTargetID: number;
  ItemStartingWeapon: any;
  ItemStartingWeaponTarget: string;
  ItemStartingWeaponTargetID: number;
  JobIndex: number;
  LimitBreak1: any;
  LimitBreak1Target: string;
  LimitBreak1TargetID: number;
  LimitBreak2: any;
  LimitBreak2Target: string;
  LimitBreak2TargetID: number;
  LimitBreak3: any;
  LimitBreak3Target: string;
  LimitBreak3TargetID: number;
  ModifierDexterity: number;
  ModifierHitPoints: number;
  ModifierIntelligence: number;
  ModifierManaPoints: number;
  ModifierMind: number;
  ModifierPiety: number;
  ModifierStrength: number;
  ModifierVitality: number;
  MonsterNote: any;
  MonsterNoteTarget: string;
  MonsterNoteTargetID: number;
  Name: string;
  NameEnglish: string;
  NameEnglish_de: string;
  NameEnglish_en: string;
  NameEnglish_fr: string;
  NameEnglish_ja: string;
  Name_de: string;
  Name_en: string;
  Name_fr: string;
  Name_ja: string;
  PartyBonus: number;
  Prerequisite: any;
  PrerequisiteTarget: string;
  PrerequisiteTargetID: number;
  PrimaryStat: number;
  PvPActionSortRow: number;
  RelicQuest: any;
  RelicQuestTarget: string;
  RelicQuestTargetID: number;
  Role: number;
  StartingLevel: number;
  StartingTown: any;
  StartingTownTarget: string;
  StartingTownTargetID: number;
  UIPriority: number;
  UnlockQuest: any;
  UnlockQuestTarget: string;
  UnlockQuestTargetID: number;
}
// ...
export interface EquipSlotCategory {
  Body: number;
  Ears: number;
  Feet: number;
  FingerL: number;
  FingerR: number;
  Gloves: number;
  Head: number;
  ID: number;
  Legs: number;
  MainHand: number;
  Neck: number;
  OffHand: number;
  SoulCrystal: number;
  Waist: number;
  Wrists: number;
}
// ...
export interface ItemRepair {
  ID: number;
  Item: Item2;
  ItemTarget: string;
  ItemTargetID: number;
}
// ...
export interface ItemSortCategory {
  ID: number;
  Param: number;
}

export interface ItemUicategory {
  ID: number;
  Icon: string;
  IconHD: string;
  IconID: number;
  Name: string;
  Name_de: string;
  Name_en: string;
  Name_fr: string;
  Name_ja: string;
  OrderMajor: number;
  OrderMinor: number;
}

export interface Title {
  ID: number;
  Icon: string;
  IsPrefix: number;
  Name: string;
  NameFemale: string;
  NameFemale_de: string;
  NameFemale_en: string;
  NameFemale_fr: string;
  NameFemale_ja: string;
  Name_de: string;
  Name_en: string;
  Name_fr: string;
  Name_ja: string;
  Order: number;
}